refactor: move split_line_at_distances to geometry.py
