+++ test/test_split.py
@@ -20,7 +20,6 @@ from shapely import (
 
 from geogenalg.split import (
     explode_and_hash_id,
-    split_line_at_distances,
     split_lines_by_points,
 )
 
@@ -544,141 +543,6 @@ def test_split_lines_by_points(
 
     for result_geom, expected_geom in zip(
         result.geometry,
-        expected_geometries,
-        strict=True,
-    ):
-        assert equals_exact(
-            result_geom,
-            expected_geom,
-            tolerance=1e-9,
-        )
-
-
-@pytest.mark.parametrize(
-    (
-        "line",
-        "distances",
-        "expected_geometries",
-    ),
-    [
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [10, 0],
-                ]
-            ),
-            [5],
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [5, 0],
-                    [10, 0],
-                ]
-            ),
-            [5],
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [10, 0],
-                ]
-            ),
-            [2, 7],
-            [
-                LineString([[0, 0], [2, 0]]),
-                LineString([[2, 0], [7, 0]]),
-                LineString([[7, 0], [10, 0]]),
-            ],
-        ),
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [5, 0],
-                    [10, 0],
-                ]
-            ),
-            [2.5, 7.5],
-            [
-                LineString([[0, 0], [2.5, 0]]),
-                LineString([[2.5, 0], [5, 0], [7.5, 0]]),
-                LineString([[7.5, 0], [10, 0]]),
-            ],
-        ),
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [10, 0],
-                ]
-            ),
-            [],
-            [
-                LineString([[0, 0], [10, 0]]),
-            ],
-        ),
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [10, 0],
-                ]
-            ),
-            [0, 10],
-            [
-                LineString([[0, 0], [10, 0]]),
-            ],
-        ),
-        (
-            LineString(
-                [
-                    [0, 0],
-                    [10, 0],
-                ]
-            ),
-            [5, 5],
-            [
-                LineString([[0, 0], [5, 0]]),
-                LineString([[5, 0], [10, 0]]),
-            ],
-        ),
-    ],
-    ids=[
-        "single split",
-        "split at existing vertex",
-        "multiple splits",
-        "multiple segments",
-        "no splits",
-        "ignore endpoints",
-        "duplicate distances",
-    ],
-)
-def test_split_line_at_distances(
-    line: LineString,
-    distances: list[float],
-    expected_geometries: list[LineString],
-):
-    result = split_line_at_distances(
-        line,
-        distances,
-    )
-
-    assert len(result) == len(expected_geometries)
-
-    for result_geom, expected_geom in zip(
-        result,
         expected_geometries,
         strict=True,
     ):
